requirements.txt
beautifulsoup4>=4.12.0
pytest>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0
discord.py>=2.0.0